Feature : 프로모션/세일 로직 분리
