+++ src/basic/main.basic.js
@@ -19,6 +19,7 @@ import {
 import { calculatePoints } from "./services/points.js";
 import { handleCalculateCartStuff } from "./core/cart.js";
 import { setupAddToCartHandler, setupCartItemHandler } from "./handlers/cartHandlers.js";
+import { startLightningSale, startSuggestedPromotion } from "./services/promotions.js";
 
 let prodList;
 let bonusPts = 0;
@@ -120,52 +121,8 @@ function main() {
   }
   onUpdateSelectOptions();
   handleCalculateCartStuffWrapper();
-  lightningDelay = Math.random() * 10000;
-  setTimeout(() => {
-    setInterval(function () {
-      let luckyIdx = Math.floor(Math.random() * prodList.length);
-      let luckyItem = prodList[luckyIdx];
-      if (luckyItem.q > 0 && !luckyItem.onSale) {
-        luckyItem.val = Math.round((luckyItem.originalVal * 80) / 100);
-        luckyItem.onSale = true;
-        alert("⚡번개세일! " + luckyItem.name + "이(가) 20% 할인 중입니다!");
-        onUpdateSelectOptions();
-        doUpdatePricesInCart();
-      }
-    }, 30000);
-  }, lightningDelay);
-  setTimeout(function () {
-    setInterval(function () {
-      if (cartDisp.children.length === 0) {
-      }
-      if (lastSel) {
-        let suggest = null;
-        for (let k = 0; k < prodList.length; k++) {
-          if (prodList[k].id !== lastSel) {
-            if (prodList[k].q > 0) {
-              if (!prodList[k].suggestSale) {
-                suggest = prodList[k];
-                break;
-              }
-            }
-          }
-        }
-        if (suggest) {
-          alert(
-            "💝 " +
-              suggest.name +
-              `은(는) 어떠세요? 지금 구매하시면 ${
-                DISCOUNT_RATES.SUGGESTED_DISCOUNT * 100
-              }% 추가 할인!`
-          );
-          suggest.val = Math.round((suggest.val * (100 - 5)) / 100);
-          suggest.suggestSale = true;
-          onUpdateSelectOptions();
-          doUpdatePricesInCart();
-        }
-      }
-    }, 60000);
-  }, Math.random() * 20000);
+  startLightningSale(prodList, onUpdateSelectOptions, doUpdatePricesInCart);
+  startSuggestedPromotion(cartDisp, prodList, lastSel, onUpdateSelectOptions, doUpdatePricesInCart);
 }
 let sum;
